Refactor to move celery directory to na_celery
