--- tests/app/celery/test_email_tasks.py
@@ -1,10 +1,0 @@
-class WhenProcessingSendEmails:
-
-    def it_sends_an_email_to_members_up_to_email_limit(self):
-        pass
-
-    def it_does_not_send_an_email_if_not_between_start_and_expiry(self):
-        pass
-
-    def it_sends_email_with_correct_template(self):
-        pass
